+++ prept/providers.py
@@ -2,12 +2,12 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, ClassVar
-from prept.errors import TemplateProviderNotFound, InvalidConfig, PreptCLIError
+from typing import TYPE_CHECKING, Any
+from prept.errors import InvalidConfig, SpecResolutionError, PreptCLIError
+from prept import utils
 
 import string
 import pathlib
-import importlib
 
 try:
     import jinja2
@@ -21,121 +21,10 @@ if TYPE_CHECKING:
     from prept.file import BoilerplateFile
 
 __all__ = (
-    'resolve_template_provider',
-    'get_prept_template_provider',
     'TemplateProvider',
     'StringTemplateProvider',
     'Jinja2TemplateProvider',
 )
-
-
-def get_prept_template_provider(name: str) -> type[TemplateProvider] | None:
-    """Prept's default template provider resolver.
-
-    Packages providing custom template providers should define
-    this function at module level. Prept will call this function
-    at generation time with the provider name from preptconfig.json.
-
-    This function should return a subclass of :class:`TemplateProvider`
-    or ``None`` if provider name is invalid.
-
-    Parameters
-    ~~~~~~~~~~
-    name: :class:`str`
-        The provider's name.
-
-    Returns
-    ~~~~~~~
-    type[:class:`TemplateProvider`]
-        The template provider resolved from given name.
-    """
-    if name == StringTemplateProvider.name:
-        return StringTemplateProvider
-
-    if name == Jinja2TemplateProvider.name:
-        if not _JINJA2_INSTALLED:
-            raise PreptCLIError(
-                'Jinja must be installed in order to use the "jinja2" template provider. ',
-                hint='See https://jinja.palletsprojects.com/en/stable/intro/#installation for help on installing Jinja2'
-            )
-
-        return Jinja2TemplateProvider
-
-    return None
-
-
-def resolve_template_provider(spec: str) -> type[TemplateProvider]:
-    """Resolves a template provider from its spec.
-
-    The spec is given in one the following format:
-
-    - ``provider-name``
-    - ``provider-class-name``
-    - ``module_name:provider-name``
-    - ``package:provider-class-name``
-
-    If no ``module_name`` is provided, it is assumed that a built-in
-    template provider from Prept is needed.
-
-    The ``provider-name`` is the :attr:`TemplateProvider.name` attribute
-    of template provider and if provided, the provider is resolved through
-    the :func:`get_prept_template_provider` function defined by the module_name.
-
-    If ``provider-name`` resolution fails, ``provider-class-name`` resolution
-    is performed.
-
-    .. versionchanged:: 0.2.0
-
-        This function now takes spec in standard Python module format i.e. ``module_name:object``
-        instead of ``module_name::object``.
-
-    Returns
-    ~~~~~~~
-    type[:class:`TemplateProvider`]
-        The resolved template provider.
-    """
-    parts = spec.split(":")
-    if not parts:
-        raise InvalidConfig('template_provider', 'Template provider name cannot be empty')
-
-    if len(parts) == 1:
-        module_name = 'prept.providers'
-        provider_name = parts[0]
-    elif len(parts) == 2:
-        module_name, provider_name = parts
-        module_name = module_name.strip()
-        provider_name = provider_name.strip()
-    else:
-        raise TemplateProviderNotFound(spec, 'too many separators')
-
-    if not module_name:
-        module_name = 'prept.providers'
-    if not provider_name:
-        raise TemplateProviderNotFound(spec, 'no provider name given')
-
-    try:
-        module = importlib.import_module(module_name)
-    except ImportError:
-        raise TemplateProviderNotFound(spec, f'failed to import {module_name}')
-
-    provider = getattr(module, provider_name.strip(), None)
-    if provider is None:
-        resolver = getattr(module, 'get_prept_template_provider', None)
-        if resolver is None:
-            raise TemplateProviderNotFound(spec, 'failed to resolve')
-        try:
-            provider = resolver(provider_name)
-        except Exception as e:
-            if isinstance(e, PreptCLIError):
-                raise
-            raise TemplateProviderNotFound(spec, f'error in resolution: {e}') from None
-
-    if provider is None:
-        raise TemplateProviderNotFound(spec, 'failed to resolve')
-    if not getattr(provider, '__prept_template_provider__', False):
-        raise TemplateProviderNotFound(spec, f'not a subclass of TemplateProvider; found {provider!r}')
-
-    return provider
 
 
 class TemplateProvider:
@@ -147,7 +36,7 @@ class TemplateProvider:
 
     All template providers, external or provided by Prept, inherit from
     this class and implement the :meth:`.process_content` and :meth:`.process_path`
-    methods. All providers must also set the :attr:`.name` class attribute.
+    methods.
 
     Prept provides the following built-in template providers:
 
@@ -156,19 +45,14 @@ class TemplateProvider:
 
     Attributes
     ~~~~~~~~~~
-    name: :class:`str`
-        Class attribute.
-
-        The name used to identify the template provider.
     settings:
         Dictionary containing settings for provider from preptconfig.json.
 
         .. versionadded:: 0.4.0
     """
-    name: ClassVar[str]
 
     # This marker is used to check if given template provider inherits
-    # from this base TemplateProvider class. Because resolve_template_provider()
+    # from this base TemplateProvider class. Because TemplateProvider._resolve()
     # uses importlib to load module from spec, issubclass() helper does not work
     # properly due to separate class objects being created.
     # See this SO question: https://stackoverflow.com/q/11461356
@@ -176,6 +60,33 @@ class TemplateProvider:
 
     def __init__(self, settings: dict[str, Any]) -> None:
         self.settings = settings
+
+    @classmethod
+    def _resolve(cls, spec: str, key: str = 'template_provider') -> type[TemplateProvider]:
+        parts = spec.split(":")
+        if not parts:
+            raise InvalidConfig(key, 'Template provider name cannot be empty')
+
+        if len(parts) == 1:
+            if parts[0] not in BUILT_IN_PROVIDERS:
+                raise InvalidConfig(key, f'No built-in provider with name {parts[0]!r} found')
+
+            provider = BUILT_IN_PROVIDERS[parts[0]]
+        else:
+            provider = utils.resolve_from_module_spec_format(spec, key)
+
+        if not getattr(provider, '__prept_template_provider__', False):
+            raise SpecResolutionError(key, f'Provider is not a subclass of TemplateProvider')
+
+        provider._on_provider_resolve()
+        return provider
+
+    @classmethod
+    def _on_provider_resolve(cls) -> None:
+        # XXX: Make this public?
+        # Currently, this is used by jinja2 provider (or any future provider requiring
+        # an external dependency) to check if the dependency is installed.
+        pass
 
     def process_path(self, path: pathlib.Path, context: GenerationContext) -> pathlib.Path:
         """"Processes the given path and replaces the.
@@ -214,10 +125,8 @@ class StringTemplateProvider(TemplateProvider):
     This uses :meth:`string.Template.safe_substitute()` to ensure that any invalid
     or missing variables are silently ignored at generation time.
 
-    This is identified by the ``stringsub`` name.
+    This can be used by setting :attr:`~Boilerplate.template_provider` to ``stringsub``
     """
-
-    name = 'stringsub'
 
     def process_path(self, path: pathlib.Path, context: GenerationContext) -> pathlib.Path:
         updated = string.Template(str(path)).safe_substitute(context.variables)
@@ -262,10 +171,16 @@ class Jinja2TemplateProvider(TemplateProvider):
 
     For more information, please refer to Jinja documentation: https://jinja.palletsprojects.com/
 
-    This is identified by the ``jinja2`` name.
+    This can be used by setting :attr:`~Boilerplate.template_provider` to ``jinja2``
     """
 
-    name = 'jinja2'
+    @classmethod
+    def _on_provider_resolve(cls) -> None:
+        if not _JINJA2_INSTALLED:
+            raise PreptCLIError(
+                'Jinja must be installed in order to use the "jinja2" template provider. ',
+                hint='See https://jinja.palletsprojects.com/en/stable/intro/#installation for help on installing Jinja2'
+            )
 
     def process_path(self, path: pathlib.Path, context: GenerationContext) -> pathlib.Path:
         # get_template_provider() checks Jinja installation so this assertion
@@ -283,3 +198,9 @@ class Jinja2TemplateProvider(TemplateProvider):
         temp = jinja2.Template(src)
         
         return temp.render(context.variables)
+
+
+BUILT_IN_PROVIDERS = {
+    'stringsub': StringTemplateProvider,
+    'jinja2': Jinja2TemplateProvider,
+}